Move autocompleteInput out of PD directory
